Minor refactoring, moving some classes to their own files
Allow the retrieve threshold to be controlled via the
spinnaker.small_retrieve system property.
