+++ SpiNNaker-front-end/src/main/java/uk/ac/manchester/spinnaker/front_end/Constants.java
@@ -19,7 +19,7 @@ package uk.ac.manchester.spinnaker.front_end;
 import static java.lang.Integer.getInteger;
 import static java.lang.Math.max;
 
-/** Miscellaneous constants. */
+/** Miscellaneous constants that can be overridden by system properties. */
 public abstract class Constants {
 	/**
 	 * The name of the system property defining the number of parallel tasks
@@ -34,9 +34,7 @@ public abstract class Constants {
 	 * If a property with this name is absent, a default is used ({@code 4}).
 	 */
 	public static final String PARALLEL_PROPERTY = "spinnaker.parallel_tasks";
-	/**
-	 * Default value of {@link #PARALLEL_SIZE}.
-	 */
+	/** Default value of {@link #PARALLEL_SIZE}. */
 	private static final int PARALLEL_DEFAULT = 4;
 	/**
 	 * The number of parallel tasks that should be executed at once. The
@@ -60,9 +58,7 @@ public abstract class Constants {
 	 * If a property with this name is absent, a default is used ({@code 7}).
 	 */
 	public static final String NEXT_MSGS_PROPERTY = "spinnaker.next_messages";
-	/**
-	 * Default value of {@link #PARALLEL_SIZE}.
-	 */
+	/** Default value of {@link #PARALLEL_SIZE}. */
 	private static final int NEXT_MSGS_DEFAULT = 7;
 	/**
 	 * The number of <em>next messages</em> that should be used in the data
@@ -71,4 +67,24 @@ public abstract class Constants {
 	public static final int NEXT_MESSAGES_COUNT =
 			// Zero or less make no sense at all
 			max(0, getInteger(NEXT_MSGS_PROPERTY, NEXT_MSGS_DEFAULT));
+
+	/**
+	 * The name of the system property defining the threshold at which retrieves
+	 * are switched over to the fast (data-speed-up-packet-gatherer based)
+	 * retrieve protocol. Below this threshold, the retrieves of data are done
+	 * via a normal SCAMP memory read.
+	 * <p>
+	 * If a property with this name is absent, a default is used
+	 * ({@code 40000}).
+	 */
+	public static final String SMALL_RETRIEVE_PROPERTY =
+			"spinnaker.small_retrieve";
+	/** Default value of {@link #SMALL_RETRIEVE_THRESHOLD}. */
+	private static final int SMALL_RETRIEVE_DEFAULT = 40000;
+	/**
+	 * Retrieves of data that is less than this many bytes are done via a normal
+	 * SCAMP memory read.
+	 */
+	public static final int SMALL_RETRIEVE_THRESHOLD =
+			max(1, getInteger(SMALL_RETRIEVE_PROPERTY, SMALL_RETRIEVE_DEFAULT));
 }
