+++ SpiNNaker-front-end/src/main/java/uk/ac/manchester/spinnaker/front_end/download/GatherDownloadConnection.java
@@ -1,0 +1,126 @@
+/*
+ * Copyright (c) 2018-2019 The University of Manchester
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package uk.ac.manchester.spinnaker.front_end.download;
+
+import static uk.ac.manchester.spinnaker.messages.Constants.SCP_SCAMP_PORT;
+
+import java.io.IOException;
+import java.net.SocketTimeoutException;
+import java.nio.ByteBuffer;
+
+import uk.ac.manchester.spinnaker.connections.SDPConnection;
+import uk.ac.manchester.spinnaker.machine.ChipLocation;
+import uk.ac.manchester.spinnaker.machine.CoreLocation;
+import uk.ac.manchester.spinnaker.machine.tags.IPTag;
+import uk.ac.manchester.spinnaker.messages.sdp.SDPMessage;
+
+/**
+ * A connection for handling the Data Speed Up protocol.
+ *
+ * @author Donal Fellows
+ */
+final class GatherDownloadConnection extends SDPConnection {
+	private long lastSend = 0L;
+	/**
+	 * Packet minimum send interval, in <em>nanoseconds</em>.
+	 */
+	private static final int INTER_SEND_INTERVAL_NS = 60000;
+	/** An empty buffer. Used so we don't try to read zero bytes. */
+	private static final ByteBuffer EMPTY_DATA = ByteBuffer.allocate(0);
+
+	/**
+	 * Create an instance.
+	 *
+	 * @param location
+	 *            Where the connection is talking to.
+	 * @param iptag
+	 *            What IPtag the Data Speed Up protocol is working on.
+	 * @throws IOException
+	 *             If anything goes wrong with socket setup.
+	 */
+	GatherDownloadConnection(ChipLocation location, IPTag iptag)
+			throws IOException {
+		super(location, iptag.getBoardAddress(), SCP_SCAMP_PORT);
+	}
+
+	private void sendMsg(SDPMessage msg) throws IOException {
+		while (System.nanoTime() < lastSend + INTER_SEND_INTERVAL_NS) {
+			Thread.yield();
+		}
+		lastSend = System.nanoTime();
+		sendSDPMessage(msg);
+	}
+
+	/**
+	 * Send a message asking the extra monitor core to read from a region of
+	 * SDRAM and send it to us (using the configured IPtag).
+	 *
+	 * @param extraMonitorCore
+	 *            The location of the monitor.
+	 * @param address
+	 *            Where to read from.
+	 * @param length
+	 *            How many bytes to read.
+	 * @throws IOException
+	 *             If message sending fails.
+	 */
+	void sendStart(CoreLocation extraMonitorCore, int address, int length)
+			throws IOException {
+		sendMsg(StartSendingMessage.create(extraMonitorCore, address,
+				length));
+	}
+
+	/**
+	 * Send a message asking the extra monitor core to ask it to resend some
+	 * data.
+	 *
+	 * @param extraMonitorCore
+	 *            The location of the monitor.
+	 * @param missingSeqs
+	 *            Description of what sequence numbers are missing.
+	 * @param numPackets
+	 *            How many resend messages will be used.
+	 * @throws IOException
+	 *             If message sending fails.
+	 */
+	void sendMissing(MissingSequenceNumbersMessage msg) throws IOException {
+		sendMsg(msg);
+	}
+
+	/**
+	 * Receive a data-speed-up payload packet.
+	 *
+	 * @param timeout
+	 *            How long to wait for the packet.
+	 * @return The packet's raw contents. On timeout, an empty buffer will
+	 *         be returned. Never returns {@code null}.
+	 * @throws IOException
+	 *             If a non-recoverable error (e.g., closed channel)
+	 *             happens.
+	 */
+	ByteBuffer getNextPacket(int timeout) throws IOException {
+		try {
+			ByteBuffer b = receive(timeout);
+			if (b == null) {
+				return EMPTY_DATA;
+			}
+			return b;
+		} catch (SocketTimeoutException ignored) {
+			return EMPTY_DATA;
+		}
+	}
+}
